Split update files into a submodule.
