+++ wptrunner/wptcommandline.py
@@ -321,10 +321,8 @@ def create_parser_update():
     parser.add_argument("--branch", action="store", type=abs_path,
                         help="Remote branch to sync against")
     parser.add_argument("--rev", action="store", help="Revision to sync to")
-    parser.add_argument("--no-check-clean", action="store_true", default=False,
-                        help="Don't check the working directory is clean before updating")
-    parser.add_argument("--patch", action="store_true",
-                        help="Create an mq patch or git commit containing the changes.")
+    parser.add_argument("--no-patch", action="store_true",
+                        help="Don't create an mq patch or git commit containing the changes.")
     parser.add_argument("--sync", dest="sync", action="store_true", default=False,
                         help="Sync the tests with the latest from upstream")
     parser.add_argument("--ignore-existing", action="store_true", help="When updating test results only consider results from the logfiles provided, not existing expectations.")
